# src/app/docs/swagger.yaml
definitions:
  handler.CreateCrudRequest:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  handler.DeleteUserResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.Response'
      message:
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      errorCode:
        type: string
      message:
        type: string
    type: object
  handler.ListUsersReponse:
    properties:
      data:
        items:
          $ref: '#/definitions/schemas.Response'
        type: array
      message:
        type: string
    type: object
  handler.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  handler.LoginResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.Response'
      message:
        type: string
    type: object
  handler.SignupRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  handler.SignupResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.SignupResponse'
      message:
        type: string
    type: object
  handler.UpdateUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  handler.UpdateUserResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.Response'
      message:
        type: string
    type: object
  schemas.Response:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      role:
        type: string
      updatedtAt:
        type: string
      username:
        type: string
    type: object
  schemas.SignupResponse:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      updatedtAt:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /user/add:
    post:
      consumes:
      - application/json
      description: Create a new non-admin user from inside the app
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateCrudRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.SignupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Add User
      tags:
      - User
  /user/delete:
    delete:
      consumes:
      - application/json
      description: Delete users from the app
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateCrudRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.DeleteUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete Users
      tags:
      - User
  /user/login:
    post:
      consumes:
      - application/json
      description: Login user based on credentials saved on database
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Login User
      tags:
      - User
  /user/signup:
    post:
      consumes:
      - application/json
      description: Create admin user
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.SignupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.SignupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Signup User
      tags:
      - User
  /user/update:
    put:
      consumes:
      - application/json
      description: Update user info
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.UpdateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update User
      tags:
      - User
  /users/count:
    get:
      consumes:
      - application/json
      description: Counts the number of users registered in the app
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ListUsersReponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Count Users
      tags:
      - Users
  /users/list:
    get:
      consumes:
      - application/json
      description: List active users in the app
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ListUsersReponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List Active Users
      tags:
      - Users
swagger: "2.0"